pyproject.toml
[tool.poetry]
name = "quake-arena-log"
version = "0.1.0"
description = "Quake 3 arena log parser"
authors = ["Thiago Simões Gonçalves <thiagosgoncalves7@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
numpy = "^1.23.0"
pytest-cov = "^3.0.0"
darker = "^1.5.0"

[tool.poetry.dev-dependencies]
black = "^21.7b0"
isort = "^5.9.3"
flake8 = "^3.9.2"
darker = "^1.3.0"
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.darker]
src = ["scripts/"]
diff = true
check = true
isort = true
lint = ["flake8"]
log_level = "INFO"
